pyproject.toml
[project]
name = "causal_meta"
version = "0.1.0"
description = "Bayesian Causal Discovery Benchmarking Framework"
authors = [{name = "Anton Lechuga", email = "anton.luis.lechuga@proton.me"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "networkx>=3.0",
    "scipy>=1.10.0",
    "matplotlib>=3.7.0",
    "hydra-core>=1.3.0",
    "gpytorch>=1.11.0",
]

[project.scripts]
causal-meta = "causal_meta.runners.pipe:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]
cluster = [
    "hydra-submitit-launcher>=1.2.0",
]
wandb = [
    "wandb>=0.15.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/causal_meta"]
include = [
    "src/causal_meta/runners/configs/**/*.yaml",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"